# Fix #23077: Fix condition Safe-NULL condition on Update Trigger