@rodrigowebjump 2019-06-28
Hi @sebastien-kathmandu  did you review the functional test that failed?

https://testing-service.magento-community.engineering/reports/magento/magento2/pull/23294/70914594-551e-495a-8f97-9a8e92625ac0/26168/Functional/allure-report-ce/index.html#suites/8187718e2fd144987673ef34ee62b41b

Regards