@seyuf 2019-07-03
Yep any update on why the test failed?
As the same test is ok on 2.2-develop (I.e [here](https://github.com/magento/magento2/pull/23079/checks?check_run_id=139312976))
But failing on 2.3-develop i.e [here](https://github.com/magento/magento2/pull/23294/checks?check_run_id=150164163)

I have this fix on prod 2.3.1 ^^. It seems to be working fine...
Anyways, it would be nice if someone could confirm that this is just not nonsense in the testing process.
@sebastien-kathmandu ?