@sebastien-kathmandu 2019-07-03
Hi @seyuf and @rodrigowebjump ,

I saw the test failing, and I don't know too why this test fail, I need to setup this functional test on my local to understand why. If anyone has an idea, I would be happy to get any clue ;-) . I'm wondering if an other part also of indexing has an issue and that would explain this test failing.

Cheers,

Sebastien