@sivaschenko 2019-08-08
@magento run all tests