@hostep 2019-08-25
It looks like Magento core devs fixed this as well very recently in [MC-19407: Number of the rows increase very fast in changelog table](https://github.com/magento/magento2/commit/eec9c26a44f3a62ef426f9ac326ff36269a77d7b)