@sivaschenko 2019-08-26
@hostep indeed, thanks for the notification. @sebastien-kathmandu thank you very much for the pull request, your contribution is appreciated! However, as you can see the issue has already been fixed so I have to close this pull request.